Prototype implementation for support TypeScript with fable-splitter
The implementation doesn't work yet because we need a way to handle the TS imports.

Indeed, when detecting an import of an existing TypeScript file we should copy the TS file without passing via Babel and also detect all the import/require from that file to copy with and so on.