Pragmatically fix test with culture specific decimal separator
In many cultures, e.g. sv-se, commas are used for decimal separator.
In an interpolated string this can normally be controlled by using
the class `FormattableString.Invariant()`.
However, we also want to test %-specifiers, but that cannot be
combined with `FormattableString.Invariant()`, only general
.NET-style specifiers work. Therefore, a simple replace of commas to
the invariant style dots is used.